Create command-line tool that dumps MA synergies (#5145)
* Allow for a fan-made milestone-award selector.

* Create command-line tool that dumps MA synergies
